arreglos

+++ sitio/src/database/migrations/20211003221003-create-product.js
@@ -14,7 +14,7 @@ module.exports = {
       },
       price: {
         allowNull: false,
-        type: Sequelize.DECIMAL(8,2)
+        type: Sequelize.INTEGER
       },
       description: {
         allowNull: false,
